requirements.txt
streamlit==1.45.1
numpy==2.0.2
opencv-python-headless==4.11.0.86
tensorflow-cpu==2.18.1
keras==3.10.0
pandas==2.3.0
plotly==6.1.2
matplotlib==3.10.3
streamlit-drawable-canvas==0.9.3
gdown==5.2.0
